requirements.txt
# SendGrid removed — now using Gmail SMTP directly
PyNaCl>=1.5.0
anthropic>=0.40.0
feedparser>=6.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
